moving steps to an inner plan
