+++ Planner.ts
@@ -1,6 +1,7 @@
 ///<reference path="World.ts"/>
 ///<reference path="Interpreter.ts"/>"/>
 ///<reference path="Searcher.ts"/>
+///<reference path="InnerWorld.ts"/>
 
 module Planner {
 
@@ -24,7 +25,6 @@ module Planner {
 
 
     export interface Result extends Interpreter.Result {plan:string[];}
-    export interface coor {row:number; col:number;}
 
 
     export function planToString(res : Result) : string {
@@ -47,7 +47,7 @@ module Planner {
             var intprt = intprts[i];
             var plan : string[] = [];
             var planner : plannerViaSearch =
-                            new plannerViaSearch(intprt, flatten(state.stacks));
+                            new plannerViaSearch(intprt, InnerWorld.flatten(state.stacks));
             if(Searcher.search(planner)) {
                 var statearm = state.arm;
                 var res : Interpreter.Literal[] = planner.getResult();
@@ -81,298 +81,16 @@ module Planner {
         return pickstack;
     }
 
-    function findObjAt(newCoor : coor, literals : Interpreter.Literal[]) : string {
-        for(var i:number = 0; i < literals.length; ++i)
-            if(('in' == literals[i].rel)
-                && (newCoor.row == +literals[i].args[0])
-                && (newCoor.col == +literals[i].args[1]))
-                return literals[i].args[2];
-        return null;
-    }
-
-    function findIn(arg : string, literals : Interpreter.Literal[]) : number {
-        for(var i:number = 0; i < literals.length; ++i)
-            if(('in' == literals[i].rel) && (arg == literals[i].args[2]))
-                return i;
-        return -1;
-    }
-
-    function reverseObjsOnTop(arg : string, literals : Interpreter.Literal[]) : string[] {
-        var pos : coor = findPos(arg, literals);
-        var res : string[] = [];
-        var n : number = 0;
-        do {
-            ++n;
-            var chg : boolean = false;
-            for(var i:number = 0; i < literals.length; ++i)
-                if(('in' == literals[i].rel)
-                   && (pos.row == +literals[i].args[0])
-                   && (pos.col + n == +literals[i].args[1]) ) {
-                   var obj = literals[i].args[2];
-                   res = Array.prototype.concat.apply([obj], res);
-                   if(findClear(obj, [], literals) != -1)
-                        return res;
-                   chg = true;
-                   break;
-                }
-            if(!chg) {
-                this.printDebugInfo('Error');
-                return res;
-            }
-        } while(true);
-    }
-
-    function findClear(arg : string, used : string[], literals : Interpreter.Literal[]) : number {
-        for(var i:number = 0; i < literals.length; ++i)
-            if('clear' == literals[i].rel) {
-                if(arg == literals[i].args[0])
-                    return i;
-                if((arg == 'floor') && (literals[i].args[0].length > 1)) {
-                    var exists : boolean = false;
-                    for(var j:number = 0; j < literals.length; ++j)
-                        if(used[j] == literals[i].args[0])
-                            exists = true;
-                    if(!exists)
-                        return i;
-                }
-            }
-        return -1;
-    }
-
-    function findPos(arg : string, literals : Interpreter.Literal[]) : coor {
-        var i : number = findIn(arg, literals);
-        if(i==-1)
-            return null;
-        return {row : +literals[i].args[0], col : +literals[i].args[1]};
-    }
-    function setPos(arg : string, literals : Interpreter.Literal[], newCoor : coor) {
-        var i : number = findIn(arg, literals);
-        if(i==-1)
-            return;
-        literals[i].args[0] = newCoor.row.toString();
-        literals[i].args[1] = newCoor.col.toString();
-    }
-
-    function hasFloor(goals : Interpreter.Literal[]) : number {
-        for(var j = 0; j < goals.length; ++j)
-            if((goals[j].rel == 'ontop') && (goals[j].args[1] == 'floor'))
-                return j;
-        return -1;
-    }
-
     function printStates(state:Interpreter.Literal[]) : void {
         state.forEach((literal) => {
         });
-    }
-
-    interface Step {
-        stepPlan : Interpreter.Literal[];
-        stepNumber() : number;
-        isPreRequisitesOk(goals : Interpreter.Literal[], state:Interpreter.Literal[], suggest:number) : boolean;
-        performStep(goals:Interpreter.Literal[], state:Interpreter.Literal[]) : void;
-    }
-
-    //////////////////////////////////////
-    ///
-    ///  Strategic Steps. Too simple right now
-
-    class basicStep1 implements Step {
-        stepPlan : Interpreter.Literal[];
-
-        floorNumber : number = -1;
-        goalToDo : number = -1;
-
-        stepNumber() : number {return 1;}
-        isPreRequisitesOk(goals : Interpreter.Literal[], state:Interpreter.Literal[], suggest:number) : boolean {
-            var floorsNeeded : number = 0;
-            this.floorNumber = hasFloor(goals);
-            if(this.floorNumber == -1)
-                return false;
-            var goal : Interpreter.Literal = goals[this.floorNumber];
-            var clearObj1 : boolean = findClear(goal.args[0], [], state) != -1;
-            var clearObj2 : boolean = findClear(goal.args[1], [], state) != -1;
-            if((!clearObj1) || (!clearObj2))
-                return false;
-            this.goalToDo = this.floorNumber;
-            return true;
-        }
-        performStep(goals : Interpreter.Literal[], state:Interpreter.Literal[]) : void {
-            var goal : Interpreter.Literal = goals[this.goalToDo];
-            goals.splice(this.goalToDo,1);
-            var i1 : coor = findPos(goal.args[0], state);
-            var floor : number = findClear(goal.args[1], [], state);
-            var floorName : string = state[floor].args[0];
-            var i2 : coor = findPos(floorName, state);
-            setPos(goal.args[0], state, {row:i2.row, col:i2.col+1});
-            state.splice(floor,1);
-            state.push({pol: true, rel: 'clear', args: [findObjAt({row:i1.row, col:i1.col-1}, state)]});
-            this.stepPlan = [];
-            this.stepPlan.push({pol: true, rel: 'move', args: [goal.args[0], i1.row.toString() ,i2.row.toString()]});
-        }
-    }
-
-    class basicStep2 implements Step {
-        stepPlan : Interpreter.Literal[];
-
-        floorNumber : number = -1;
-        goalToDo : number = -1;
-
-        floorName : string;
-        tempName : string;
-
-        stepNumber() : number {return 1;}
-        isPreRequisitesOk(goals : Interpreter.Literal[], state:Interpreter.Literal[], suggest:number) : boolean {
-            var floorsNeeded : number = 0;
-            this.floorNumber = hasFloor(goals);
-            if(this.floorNumber == -1)
-                return false;
-            var goal : Interpreter.Literal = goals[this.floorNumber];
-            var floor : number = findClear(goal.args[1], [], state);
-            if(floor == -1)
-                return false;
-            this.floorName = state[floor].args[0];
-            var tempFloor : number = findClear(goal.args[1], [this.floorName], state);
-            if(tempFloor == -1)
-                return false;
-            this.tempName = state[tempFloor].args[0];
-            this.goalToDo = this.floorNumber;
-            return true;
-        }
-        performStep(goals : Interpreter.Literal[], state:Interpreter.Literal[]) : void {
-            var goal : Interpreter.Literal = goals[this.goalToDo];
-            var objsOntop : string[] = reverseObjsOnTop(goal.args[0], state);
-            var itemp : coor = findPos(this.tempName, state);
-            var col : number = 0;
-            var prev : string;
-            this.stepPlan = [];
-            objsOntop.forEach((obj) => {
-                ++col;
-                prev = obj;
-                var i1 : coor = findPos(obj, state);
-                setPos(obj, state, {row:itemp.row, col:col});
-                this.stepPlan.push({pol: true, rel: 'move', args: [obj, i1.row.toString() ,itemp.row.toString()]});
-            });
-            state.splice(findClear(objsOntop[0], [], state),1);
-            state.splice(findClear(this.tempName, [], state),1);
-            state.push({pol: true, rel: 'clear', args: [prev]});
-            state.push({pol: true, rel: 'clear', args: [goal.args[0]]});
-            var chainedStep : Step = new basicStep1();
-            var tst : boolean = chainedStep.isPreRequisitesOk(goals, state, this.goalToDo);
-
-            chainedStep.performStep(goals, state);
-            chainedStep.stepPlan.forEach((step) => {
-                this.stepPlan.push(step);
-            });
-        }
-    }
-
-    class basicStep3 implements Step {
-        stepPlan : Interpreter.Literal[];
-
-        tempName : string = null;
-        goalToDo : number = -1;
-
-        stepNumber() : number {return 1;}
-        isPreRequisitesOk(goals : Interpreter.Literal[], state:Interpreter.Literal[], suggest:number) : boolean {
-            var goal : Interpreter.Literal = goals[suggest];
-            var clearObj2 : boolean = findClear(goal.args[1], [], state) != -1;
-            if(!clearObj2)
-                return false;
-            var clearObj1 : boolean = findClear(goal.args[0], [], state) != -1;
-            if(!clearObj1) {// need a floor
-                var floor : number = findClear('floor', [], state);
-                if(floor == -1)
-                    return false;
-                this.tempName = state[floor].args[0];
-            }
-            this.goalToDo = suggest;
-            return true;
-        }
-        performStep(goals : Interpreter.Literal[], state:Interpreter.Literal[]) : void {
-            var goal : Interpreter.Literal = goals[this.goalToDo];
-            goals.splice(this.goalToDo,1);
-            this.stepPlan = [];
-            if(this.tempName != null) {
-                var col : number = 0;
-                var prev : string;
-                var itemp : coor = findPos(this.tempName, state);
-                var objsOntop : string[] = reverseObjsOnTop(goal.args[0], state);
-                objsOntop.forEach((obj) => {
-                    ++col;
-                    prev = obj;
-                    var i1 : coor = findPos(obj, state);
-                    setPos(obj, state, {row:itemp.row, col:col});
-                    this.stepPlan.push({pol: true, rel: 'move', args: [obj, i1.row.toString() ,itemp.row.toString()]});
-                });
-                state.splice(findClear(objsOntop[0], [], state),1);
-                state.splice(findClear(this.tempName, [], state),1);
-                state.push({pol: true, rel: 'clear', args: [prev]});
-                state.push({pol: true, rel: 'clear', args: [goal.args[0]]});
-            }
-            var i1 : coor = findPos(goal.args[0], state);
-            var support : number = findClear(goal.args[1], [], state);
-            var supportName : string = state[support].args[0];
-            var i2 : coor = findPos(supportName, state);
-            setPos(goal.args[0], state, {row:i2.row, col:i2.col+1});
-            state.splice(support,1);
-            state.push({pol: true, rel: 'clear', args: [findObjAt({row:i1.row, col:i1.col-1}, state)]});
-            this.stepPlan.push({pol: true, rel: 'move', args: [goal.args[0], i1.row.toString() ,i2.row.toString()]});
-        }
-    }
-
-    class basicStep4 implements Step {
-        stepPlan : Interpreter.Literal[];
-
-        tempName : string = null;
-        goalToDo : number = -1;
-
-        stepNumber() : number {return 1;}
-        isPreRequisitesOk(goals : Interpreter.Literal[], state:Interpreter.Literal[], suggest:number) : boolean {
-            var goal : Interpreter.Literal = goals[suggest];
-            {// need a floor
-                var floor : number = findClear('floor', [], state);
-                if(floor == -1)
-                    return false;
-                this.tempName = state[floor].args[0];
-            }
-            this.goalToDo = suggest;
-            return true;
-        }
-        performStep(goals : Interpreter.Literal[], state:Interpreter.Literal[]) : void {
-            var goal : Interpreter.Literal = goals[this.goalToDo];
-            this.stepPlan = [];
-            {
-                var col : number = 0;
-                var prev : string;
-                var itemp : coor = findPos(this.tempName, state);
-                var objsOntop : string[] = reverseObjsOnTop(goal.args[1], state);
-                objsOntop.forEach((obj) => {
-                    ++col;
-                    prev = obj;
-                    var i1 : coor = findPos(obj, state);
-                    setPos(obj, state, {row:itemp.row, col:col});
-                    this.stepPlan.push({pol: true, rel: 'move', args: [obj, i1.row.toString() ,itemp.row.toString()]});
-                });
-                state.splice(findClear(objsOntop[0], [], state), 1);
-                state.splice(findClear(this.tempName, [], state), 1);
-                state.push({pol: true, rel: 'clear', args: [prev]});
-                state.push({pol: true, rel: 'clear', args: [goal.args[1]]});
-            }
-            var chainedStep : Step = new basicStep3();
-            var tst : boolean = chainedStep.isPreRequisitesOk(goals, state, this.goalToDo);
-
-            chainedStep.performStep(goals, state);
-            chainedStep.stepPlan.forEach((step) => {
-                this.stepPlan.push(step);
-            });
-        }
     }
 
     //////////////////////////////////////
     ///
     ///  State follow up
 
-    interface mneumonicMeaning {plan : Step[];
+    interface mneumonicMeaning {plan : InnerWorld.Step[];
                                 state : Interpreter.Literal[];
                                 todo : Interpreter.Literal[];
                                 goalNumber : number;}
@@ -403,7 +121,7 @@ module Planner {
             return ret;
         }
 
-        private currentPlan : Step[];
+        private currentPlan : InnerWorld.Step[];
 
         private currentState : Interpreter.Literal[];
         private todo : Interpreter.Literal[];
@@ -431,7 +149,7 @@ module Planner {
             this.goalNumber = this.mneumonicCollection[mne].goalNumber;
         }
 
-        getCostOfCurrentState(): number {
+        getHeuristicCostOfCurrentState(): number {
             var cost:number = 10000;
             var symbols : collections.Set<string> = new collections.Set<string>();
             this.todo.forEach((op) => {
@@ -460,7 +178,7 @@ module Planner {
             return false;
         }
         equalGoalStates(A : Interpreter.Literal[], B : Interpreter.Literal[]) : boolean {
-            var tests = {ontop: this.ontop};
+            var tests = {ontop: InnerWorld.ontop};
             for(var j:number = 0; j < A.length; ++j) {
                 var a = tests[A[j].rel.trim()];
                 if((a != null) && (!a(A[j].args, B)))
@@ -491,28 +209,18 @@ module Planner {
             return false;
         }
 
-        ontop(args:string[], literals : Interpreter.Literal[]): boolean {
-            var pos1 : coor = findPos(args[0], literals);
-            var pos2 : coor = findPos(args[1], literals);
-            if((pos1 == null) || (pos2 == null))
-                return false;
-            if((pos1.row == pos2.row) && (pos1.col > pos2.col))
-                return true;
-            return false;
-        }
-
         nextChildAndMakeCurrent(): Boolean {
             return this.nextSiblingAndMakeCurrent();
         }
         nextSiblingAndMakeCurrent(): Boolean {
-            var a : Step;
+            var a : InnerWorld.Step;
             if(this.todo.length>0)
                 for(var n:number = 0; n<4; ++n) {
                     switch(n) {
-                        case 0:a = new basicStep1(); break;
-                        case 1:a = new basicStep2(); break;
-                        case 2:a = new basicStep3(); break;
-                        case 3:a = new basicStep4(); break;
+                        case 0:a = new InnerWorld.basicStep1(); break;
+                        case 1:a = new InnerWorld.basicStep2(); break;
+                        case 2:a = new InnerWorld.basicStep3(); break;
+                        case 3:a = new InnerWorld.basicStep4(); break;
                     }
                     if(a.isPreRequisitesOk(this.todo, this.currentState, 0)) {
                         a.performStep(this.todo, this.currentState);
@@ -556,24 +264,6 @@ module Planner {
         }
 
         printDebugInfo(info : string) : void {console.log(info);}
-    }
-
-    function flatten(stacks: string[][]) : Interpreter.Literal[] {
-        var oneIntprt : Interpreter.Literal[] = [];
-        var row : number = 0;
-        stacks.forEach((stack) => {
-            var prev : string = 'floor'+row;
-            var col : number = 0;
-            stack.forEach((ele) => {
-                oneIntprt.push({pol: true, rel: 'in', args: [row.toString(), col.toString(), prev]});
-                col++;
-                prev = ele;
-            });
-            oneIntprt.push({pol: true, rel: 'in', args: [row.toString(), col.toString(), prev]});
-            oneIntprt.push({pol: true, rel: 'clear', args: [prev]});
-            ++row;
-        });
-        return oneIntprt;
     }
 
     function getRandomInt(max) {
